pyproject.toml
[project]
name = "lfi-workflow-template"
version = "0.1.0"
description = "Two step local inocation workflow template"
authors = [
    { name = "Resonate HQ", email = "contact@resonatehq.io" }
]
dependencies = [
    "resonate-sdk==0.4.12",
]
readme = "README.md"
requires-python = "== 3.12.*"

[project.scripts]
    "app" = "src.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
managed = true
dev-dependencies = [
    "mypy==1.14.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.rye.scripts]
check-types = "dmypy run --timeout 60 -- src"
